+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-## Version 0.1.1 2025-03-28
+## Version `0.1.2` 2025-03-28
+### Changed
+- Fixed import error introduced in `0.1.2`.
+
+## Version `0.1.1` 2025-03-28
 ### Changed
 - Minimum supported python version changed from 3.12 to 3.10, so more users can use this library!
 - `numpy`, `pandas`, and `torch` are now optional dependencies so adding `crab-dbg` won't blow up your dependencies.
 
-## Version 0.1.0 2024-10-08
+## Version `0.1.0` 2024-10-08
 ### Added
 - Python equivalent to Rust's `dbg!()` macro.